006--Move some components outside of their folders
