+++ src/app/shared/lawnmower-details/lawnmower-details.component.scss
